requirements.txt
fastapi
uvicorn
pandas
vaderSentiment
pyjwt
python-multipart
transformers